AL2023 6.1: Update lustrefsx to 2.15.6-fsx17 commit

+++ drivers/staging/lustrefsx/lustre/mdc/mdc_locks.c
@@ -119,7 +119,7 @@ int mdc_set_lock_data(struct obd_export *exp, const struct lustre_handle *lockh,
 		LASSERTF(old_inode->i_state & I_FREEING,
 			 "Found existing inode %p/%lu/%u state %lu in lock: setting data to %p/%lu/%u\n",
 			 old_inode, old_inode->i_ino, old_inode->i_generation,
-			 old_inode->i_state,
+			 (unsigned long)old_inode->i_state,
 			 new_inode, new_inode->i_ino, new_inode->i_generation);
 	}
 	lock->l_resource->lr_lvb_inode = new_inode;
